worked on styling signin/out

+++ app.js
@@ -22,30 +22,60 @@ app.use(
 );
 
 app.get("/", function(req, resp) {
-  db.query("SELECT * FROM task").then(function(results) {
+  db.query("SELECT * FROM task WHERE userid = ($1)", userid).then(function(
+    results
+  ) {
     resp.render("todolist.html", {
       results: results
     });
   });
 });
 
+app.get("/signin", function(req, res, next) {
+  res.render("signin.html");
+});
+
+app.get("/signup", function(req, res, next) {
+  res.render("signup.html");
+});
+
 app.post("/", function(req, resp, next) {
   var id = req.body.id;
-  db.query("UPDATE task SET done = true WHERE id = ($1)", id);
+  db.query(
+    "UPDATE task SET done = true WHERE id = ($1) userid = ($2)",
+    id,
+    userid
+  );
   resp.redirect("/");
 });
 
 app.post("/add", function(req, resp, next) {
   var task = req.body.task;
   db.query(
-    "INSERT INTO task (id, description, done) VALUES (default, $1, false)",
+    "INSERT INTO task (userid, id, description, done) VALUES ($1, default, $2, false)",
+    userid,
     task
   );
   resp.redirect("/");
 });
 
+app.post("/new_user", function(req, resp, next) {
+  var email = req.body.inputEmail;
+  var password = req.body.inputPassword;
+  var fname = req.body.inputFName;
+  var lname = req.body.inputLName;
+  console.log(email, password, lname, fname);
+  db.query(
+    "INSERT INTO users (userid, email, password, last_name, first_name) VALUES (default, $1, $2, $3, $4)",
+    [email, password, lname, fname]
+  );
+  resp.redirect("/");
+});
+
 app.get("/todos/done", function(req, resp) {
-  db.query("SELECT * FROM task").then(function(results) {
+  db.query("SELECT * FROM task WHERE userid = ($1)", userid).then(function(
+    results
+  ) {
     resp.render("done.html", {
       results: results
     });
@@ -54,16 +84,20 @@ app.get("/todos/done", function(req, resp) {
 
 app.post("/todos/done", function(req, resp, next) {
   var id = req.body.id;
-  db.query("UPDATE task SET done = false WHERE id = ($1)", id);
+  db.query(
+    "UPDATE task SET done = false WHERE id = ($1) userid = ($2)",
+    id,
+    userid
+  );
   resp.redirect("/todos/done");
 });
 
 app.post("/todos/delete", function(req, resp, next) {
   var id = req.body.id;
-  db.query("DELETE FROM task WHERE id = ($1)", id);
+  db.query("DELETE FROM task WHERE id = ($1) userid = ($2)", id, userid);
   resp.redirect("/todos/done");
 });
 
-app.listen(process.env.PORT || 8000, function() {
-  console.log("Listening on port 8000");
+app.listen(process.env.PORT || 9000, function() {
+  console.log("Listening on port 9000");
 });
